Factor out TotalDetails
